导游交账总收入，实际收入的金额调整

+++ ctm_project/apply/src/main/java/eqlee/ctm/apply/entry/entity/bo/ApplyDoExaInfo.java
@@ -60,7 +60,7 @@ public class ApplyDoExaInfo {
    /**
     * 支付金额
     */
-   private Integer payPrice;
+   private Double payPrice;
 
    //同行   出行日期  线路名  联系人  电话  保名人数  结算金额
